Adjust wrong code cast

+++ src/backward.h
@@ -3057,7 +3057,7 @@ public:
         if (st.size() == 0) {
             return;
         }
-        _symbols.reset(backtrace_symbols(st.begin(), st.size()));
+        _symbols.reset(backtrace_symbols(st.begin(), (int) st.size()));
     }
 
     ResolvedTrace resolve(ResolvedTrace trace)
